+++ src/knowlang/chat_bot/chat_graph.py
@@ -1,23 +1,25 @@
 # __future__ annotations is necessary for the type hints to work in this file
 from __future__ import annotations
+
 from dataclasses import dataclass
-from typing import AsyncGenerator, List, Dict, Any, Optional
-from pydantic import BaseModel
-from pydantic_graph import BaseNode, EndStep, Graph, GraphRunContext, End, HistoryStep
-from pydantic_ai import Agent
-import logfire
 from enum import Enum
-from rich.console import Console
+from typing import Any, AsyncGenerator, Dict, List, Optional
+
+import logfire
 import voyageai
+from pydantic import BaseModel
+from pydantic_ai import Agent
+from pydantic_graph import (BaseNode, End, EndStep, Graph, GraphRunContext,
+                            HistoryStep)
+from rich.console import Console
 from voyageai.object.reranking import RerankingObject
 
-from knowlang.configs.config import AppConfig, RerankerConfig, EmbeddingConfig
-from knowlang.vector_stores.base import SearchResult, VectorStore
+from knowlang.configs.config import AppConfig, EmbeddingConfig, RerankerConfig
+from knowlang.models import EmbeddingInputType, generate_embedding
+from knowlang.utils.chunking_util import truncate_chunk
 from knowlang.utils.fancy_log import FancyLogger
 from knowlang.utils.model_provider import create_pydantic_model
-from knowlang.utils.chunking_util import truncate_chunk
-from knowlang.models.embeddings import EmbeddingInputType, generate_embedding
-
+from knowlang.vector_stores.base import SearchResult, VectorStore
 
 LOG = FancyLogger(__name__)
 console = Console()
