+++ src/knowlang/indexing/chunk_indexer.py
@@ -1,12 +1,12 @@
-from typing import List, Set
 from pathlib import Path
+from typing import List, Set
 
+from knowlang.configs.config import AppConfig
 from knowlang.core.types import CodeChunk, DatabaseChunkMetadata
 from knowlang.indexing.indexing_agent import IndexingAgent
-from knowlang.models.embeddings import generate_embedding
-from knowlang.vector_stores.base import VectorStore
-from knowlang.configs.config import AppConfig
+from knowlang.models import generate_embedding
 from knowlang.utils.fancy_log import FancyLogger
+from knowlang.vector_stores.base import VectorStore
 from knowlang.vector_stores.factory import VectorStoreFactory
 
 LOG = FancyLogger(__name__)
