+++ src/knowlang/indexing/indexing_agent.py
@@ -1,13 +1,14 @@
 from typing import List
+
 from pydantic_ai import Agent
 from rich.progress import Progress
 
 from knowlang.configs.config import AppConfig
 from knowlang.core.types import CodeChunk, DatabaseChunkMetadata
+from knowlang.models import generate_embedding
 from knowlang.utils.chunking_util import format_code_summary
 from knowlang.utils.fancy_log import FancyLogger
 from knowlang.utils.model_provider import create_pydantic_model
-from knowlang.models.embeddings import generate_embedding
 from knowlang.vector_stores.factory import VectorStoreFactory
 
 LOG = FancyLogger(__name__)
