+++ src/knowlang/models/__init__.py
@@ -1,0 +1,8 @@
+from .embeddings import generate_embedding
+from .types import EmbeddingInputType, EmbeddingVector
+
+__all__ = [
+    "EmbeddingInputType",
+    "generate_embedding",
+    "EmbeddingVector",
+]
