+++ src/knowlang/utils/migration/voyage_embedding_migrations.py
@@ -1,12 +1,14 @@
-from pathlib import Path
 import asyncio
+from pathlib import Path
+from typing import List
+
 import chromadb
 from chromadb.errors import InvalidCollectionException
-from rich.progress import Progress
 from rich.console import Console
-from typing import List
+from rich.progress import Progress
+
 from knowlang.configs.config import AppConfig, EmbeddingConfig
-from knowlang.models.embeddings import generate_embedding, EmbeddingInputType
+from knowlang.models import EmbeddingInputType, generate_embedding
 from knowlang.utils.fancy_log import FancyLogger
 
 LOG = FancyLogger(__name__)
